Able to put instructor service in a different version

+++ instructor-service/src/main/java/paubel/cs5012/student/InstructorController.java
@@ -9,36 +9,52 @@ package paubel.cs5012.instructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Iterator;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/instructor")
 public class InstructorController {
-    // Student repository object
+
+    @Autowired
+    SourceBean source;
+
+    // Instructor repository object
     private final InstructorRepository instructorRepository;
 
-    // Inject the student repository object
+    // Inject the instructor repository object
     @Autowired
     public InstructorController(InstructorRepository instructorRepository) {
         this.instructorRepository = instructorRepository;
     }
 
-    // Get all students in the database
+    // Get all instructors in the database
     @GetMapping
     public List<Instructor> getInstructors() {
         return instructorRepository.findAll();
     }
 
-    // Insert a student into the database
+    // Insert an instructor into the database
     @PostMapping
     public Instructor addInstructor(@RequestBody Instructor instructor) {
+        source.publishInstructorChange("POST", instructor.getName());
         return instructorRepository.save(instructor);
     }
 
-    // Delete a student in the database by ID
+    // Delete an instructor in the database by ID
     @DeleteMapping("/{id}")
     public void deleteInstructor(@PathVariable Long id) {
+        Optional<Instructor> instructor = instructorRepository.findById(id);
+
+        if (!instructor.isPresent()) {
+            System.out.println("Attempted to DELETE: " + instructor.get().getName() + ", but instructor does not exist!");
+            return;
+        }
+
+        source.publishInstructorChange("DELETE", instructor.get().getName());
         instructorRepository.deleteById(id);
     }
 }
